type rearrangement
